+++ libraries/SoftwareSerial/SoftwareSerial.cpp
@@ -440,7 +440,7 @@ int SoftwareSerial::available()
   return (_receive_buffer_tail + _SS_MAX_RX_BUFF - _receive_buffer_head) % _SS_MAX_RX_BUFF;
 }
 
-long SoftwareSerial::write(uint8_t b)
+ssize_t SoftwareSerial::write(uint8_t b)
 {
   if (_tx_delay == 0)
     return -1;
